requirements.txt
flask
gunicorn
pillow